// src/app/portfolio/interfaces/technologies.interface.ts
export interface TechRefs {
  frontLanguageIds?: number[]; // p.ej. TypeScript(1), JavaScript(2), Dart(5)
  backLanguageIds?: number[]; // p.ej. PHP(4), NodeJs(2), Java(1)
  frontFrameworkIds?: number[]; // p.ej. Angular(1), React(2), Flutter(3)
  backFrameworkIds?: number[]; // p.ej. Laravel(1)
  uiLibraryIds?: number[]; // p.ej. PrimeNG(1), daisyUI(2)
  versionControlToolIds?: number[]; // p.ej. Git(1), GitHub(2)
  databaseIds?: number[]; // PostgreSQL(1), MySQL(2), MongoDB(3)
  embeddedDatabaseIds?: number[]; // Isar(1)...
  clientStorageIds?: number[]; // LocalStorage(1)...
  baasPlatformIds?: number[]; // Firebase(1)...
  cloudPlatformIds?: number[]; // Cloudflare(1)...
  methodologyIds?: number[]; // Clean Architecture(23), SOLID(26), DRY(29)...
  dataTechIds?: number[]; // SQL(26), etc.
}